🛠️ model.py -> Removed type casting in operations 🛠️ train.py -> Increased BATCH_SIZE, changed data type

+++ model.py
@@ -1,17 +1,17 @@
 import jax
 
 def feed_forward(params, x):
-    x = jax.numpy.dot(x.astype(jax.numpy.bfloat16), params['in_weight'].astype(jax.numpy.bfloat16))
+    x = jax.numpy.dot(x, params['in_weight'])
     x = jax.nn.gelu(x, approximate=True)
-    x = jax.numpy.dot(x, params['out_weight'].astype(jax.numpy.bfloat16))
-    return x.astype(jax.numpy.float32)
+    x = jax.numpy.dot(x, params['out_weight'])
+    return x
 
 def attention(params, x, mask, n_heads, scale):
     batch_size, seq_len, _ = x.shape
     # Linear transformations
-    q = jax.numpy.dot(x.astype(jax.numpy.bfloat16), params['q_linear'].astype(jax.numpy.bfloat16)).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3).astype(jax.numpy.float32)
-    k = jax.numpy.dot(x.astype(jax.numpy.bfloat16), params['k_linear'].astype(jax.numpy.bfloat16)).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3).astype(jax.numpy.float32)
-    v = jax.numpy.dot(x.astype(jax.numpy.bfloat16), params['v_linear'].astype(jax.numpy.bfloat16)).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3).astype(jax.numpy.float32)
+    q = jax.numpy.dot(x, params['q_linear']).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3)
+    k = jax.numpy.dot(x, params['k_linear']).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3)
+    v = jax.numpy.dot(x, params['v_linear']).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3)
     # Compute attention scores
     scores = jax.numpy.matmul(q, k.transpose(0, 1, 3, 2)) * scale
     scores = jax.numpy.where(mask[:seq_len, :seq_len], -jax.numpy.inf, scores)
@@ -19,8 +19,8 @@ def attention(params, x, mask, n_heads, scale):
     # Compute output
     output = jax.numpy.matmul(scores, v)
     output = output.transpose(0, 2, 1, 3).reshape(batch_size, seq_len, -1)
-    output = jax.numpy.dot(output.astype(jax.numpy.bfloat16), params['o_linear'].astype(jax.numpy.bfloat16))
-    return output.astype(jax.numpy.float32)
+    output = jax.numpy.dot(output, params['o_linear'])
+    return output
 
 def transformer_block(params, x, mask, n_heads, scale):
     r = attention(params['attention'], layer_norm(x, params['attention_norm_scale'], params['attention_norm_bias']), mask, n_heads, scale)
